Improved Dataset and updated notebooks
Improved Dataset by loading only the `patient_id` and its number of slices. Then, the load of the NIfTI file and creation of the mask is done in `__getitem__`. This will help to try different batch sizes because it does not need to have all the slices in memory.

The notebooks have been updated so they use the new version of the Dataset.

Extra: added new experiments using `batch_size=16`